refactor: adjust type strcture
