+++ __tests__/scopedLock.test.ts
@@ -1,5 +1,5 @@
 import { Mutex, ScopedLock } from "../src";
-import { MutexInterface } from "../src/mutex";
+import { MutexInterface } from "../src/types";
 import { sleepFor } from "./support/util";
 
 describe("UniqueLock with Mutex", () => {
